apiKeyChecks complete

+++ Api/Api/Api/Services/EntryServices/EntryService.cs
@@ -22,13 +22,14 @@ namespace Api.Services.EntryServices
             _userRepository = userRepository;
             _mapper = mapper;
         }
-        public async Task<EntryDto> SetEntry(AddEntry entryDto)
+        public async Task<EntryDto> SetEntry(AddEntry entryDto, string apiKey)
         {
-            var user = await _userRepository.Get(entryDto.UserName);
+            
 
-            if (user == null)
+            var userToCheck = await _userRepository.GetByApiKey(apiKey);
+            if (userToCheck == null)
             {
-                throw new UnauthorizedAccessException();    
+                throw new UnauthorizedAccessException("Du får inte redigera det här inlägget");
             }
 
             var POI = await _poiRepository.Get(entryDto.POI.Longitude, entryDto.POI.Latitude, entryDto.POI.Name);
@@ -38,7 +39,7 @@ namespace Api.Services.EntryServices
                 POI = await _poiRepository.Set(entryDto.POI);
             }
 
-            var userEntryCheck = await _entryRepository.Get(POI.Id, user.Id);
+            var userEntryCheck = await _entryRepository.Get(POI.Id, userToCheck.Id);
             if (userEntryCheck != null)
             {
                 throw new Exception("Du har redan gjort ett inlägg för den här platsen");
@@ -48,11 +49,11 @@ namespace Api.Services.EntryServices
 
             entry.POIID = POI.Id;
 
-            entry.UserId = user.Id;
+            entry.UserId = userToCheck.Id;
 
            await _entryRepository.Set(entry);
 
-           return new EntryDto() { POI = POI.Name, UserName = user.Username, Description = entryDto.Description, Rating = entryDto.Rating };
+           return new EntryDto() { POI = POI.Name, UserName = userToCheck.Username, Description = entryDto.Description, Rating = entryDto.Rating };
         }
 
 
@@ -65,7 +66,7 @@ namespace Api.Services.EntryServices
                 throw new NotFoundException("Kunde inte hitta inlägget");
             }
 
-            var userToCheck = await ApiCheck.ApiKeyCheck(apiKey); 
+            var userToCheck = await _userRepository.GetByApiKey(apiKey); 
             if(entry.UserId != userToCheck.Id)
             {
                 throw new UnauthorizedAccessException("Du får inte redigera det här inlägget");
@@ -78,7 +79,7 @@ namespace Api.Services.EntryServices
             return _mapper.Map<EntryDto>(updatedEntry);
         }
 
-        public async Task<EntryDto> DeleteEntry(long id)
+        public async Task<EntryDto> DeleteEntry(long id, string apiKey)
         {
 
             var entryDelete = await _entryRepository.GetWithTracking(id);
@@ -87,23 +88,15 @@ namespace Api.Services.EntryServices
             {
                 throw new NotFoundException("Kunde inte hitta inlägget");
             }
-
-
-            //TODO: API nyckel fix!!
-            //if(entry.UserId != 1)
-            //{
-            //    throw new UnauthorizedAccessException();
-            //}
+            var userToCheck = await _userRepository.GetByApiKey(apiKey);
+            if (entryDelete.UserId != userToCheck.Id)
+            {
+                throw new UnauthorizedAccessException("Du får inte ta bort det här inlägget");
+            }
 
             await _entryRepository.DeleteEntry(entryDelete);
 
             return _mapper.Map<EntryDto>(entryDelete);
-
-
-
         }
-
-
-
     }
 }
